Created New Scene w/ Core Loop Skeleton
Combined the AI Scene and Player-Character Scene into one in order to showcase what will be a part of the core gameplay loop. Players can move the penguin, with red spheres around representing their weapons, and fight the patrolling AI square (future Skua enemy). Hitting the square with a sphere causes it to retreat, getting hit by it causes it to lift you away and off the screen before the scene restarts. This is a necessary step in combining the different elements of the game together.